Restructure of the folder skeleton of the project
