+++ gptube/config/main.go
@@ -1,4 +1,4 @@
-package environment
+package config
 
 import (
 	"log"
@@ -9,7 +9,7 @@ import (
 
 // use godot package to load/read the .env file and
 // return the value of the key
-func Getenv(key string) string {
+func Config(key string) string {
 	// load .env file
 	err := godotenv.Load(".env")
