+++ gptube/handler/main.go
@@ -1,8 +1,6 @@
-package routes
+package handler
 
-import (
-	"github.com/gofiber/fiber/v2"
-)
+import "github.com/gofiber/fiber/v2"
 
 func HomeHandler(c *fiber.Ctx) error {
 	return c.JSON(fiber.Map{
