+++ gptube/templates/main.go
@@ -1,0 +1,1 @@
+package templates
