+++ gptube/templates/youtube_email_success.gotmpl
