--- gptube/web/main.go
@@ -1,1 +1,0 @@
-package web
