--- gptube/web/youtube/main.go
@@ -1,17 +1,0 @@
-package web
-
-import (
-	"net/smtp"
-	"os"
-)
-
-var emailAuth smtp.Auth
-
-func init() {
-	emailAuth = smtp.PlainAuth(
-		"",
-		os.Getenv("EMAIL_USER"),
-		os.Getenv("EMAIL_PASSWORD"),
-		"smtp.gmail.com",
-	)
-}
